Branched out meta_yoda.h
